requirements.txt
alembic==1.10.4
blinker==1.6.2
cli-helpers==2.3.0
click==8.1.3
configobj==5.0.8
Flask==2.3.2
Flask-Cors==3.0.10
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
prompt-toolkit==3.0.38
Pygments==2.15.1
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.12
sqlparse==0.4.4
tabulate==0.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
Werkzeug==2.3.4
